chore(processors.aws_ec2): Rename processor to adhere to the standard directory structure
