+++ plugins/processors/aws_ec2/sample.conf
